[VL] Fix `RetryOnOomMemoryTarget` only spills one single consumer on retrying
